requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
rank-bm25
nltk
numpy
pydantic
ctransformers
gradio
regex
